+++ src/components/BodyComponents.cpp
@@ -45,7 +45,8 @@ void BodyComponents::allocate(uint32 nbComponentsToAllocate) {
     assert(nbComponentsToAllocate > mNbAllocatedComponents);
 
     // Make sure capacity is an integral multiple of alignment
-    nbComponentsToAllocate = std::ceil(nbComponentsToAllocate / float(GLOBAL_ALIGNMENT)) * GLOBAL_ALIGNMENT;
+    nbComponentsToAllocate =
+        static_cast<uint32>(std::ceil(nbComponentsToAllocate / float(GLOBAL_ALIGNMENT)) * GLOBAL_ALIGNMENT);
 
     // Size for the data of a single component (in bytes)
     const size_t totalSizeBytes = nbComponentsToAllocate * mComponentDataSize + mAlignmentMarginSize;
